machine-learning/chatbot-rain-main/chatbot-rain-main/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
httpx
groq
python-dotenv
